feature(slices): split test based on history of run durations
This helps in able to split test between machines, to make a run of
a big suite of long running tests run in parallel.
it use the run history of the test from elasticsearch to create
"slice" files, with the specific tests, to fit a goal of how long
you want you run suite to be. (assuming you have enough capability
of spining as much machines you need, for example using jenkins
togther with AWS spot fleets)