feat: allow configuring post to es
es is not always available in whichever environment our test suite
runs. in the cases where it isn't available, we should be able to
configure whether or not to post the reports.

disabling the plugin altogether with `-p no:elk-reporter` causes
problems if clients of the plugin have taken advantage of the
`elk_reporter` fixture. they will get a fixture lookup error.

this allows the plugin to be loaded which won't break their
fixtures or tests that depend on it.